move up down

+++ commands/create_transcript_sem_table.php
@@ -5,7 +5,7 @@ include "config/dbc.php";
 $create = "CREATE TABLE transcript_sem (
     `id` INTEGER UNSIGNED PRIMARY KEY AUTO_INCREMENT,
     `idnum` INTEGER UNSIGNED NOT NULL,
-    `ordinal` INTEGER(3) NOT NULL,
+    `ordinal` INTEGER NOT NULL,
     `sy` VARCHAR(30) NOT NULL,
     `school` VARCHAR(101) NOT NULL,
     `address` VARCHAR(101) NOT NULL,
